CUSTOM EVENTS CAN ACCEPT ARGUMENTS
we can use pass arguments with our custom events to make the parent component change values.

Let’s say we want to modify the changeUsername event so that we can pass it a value.

The $emit method takes an optional second parameter for arguments. So all we do is add our new username value after the name of our event.

this.$emit('changeUsername', 'mattmaribojoc')
Copy
Then, in our parent component, we can either access these values inline by using a special $event variable, or we can write a handler method that takes a parameter.

<account-info :username="user.username" @changeUsername="user.username = $event"/>

OR

<account-info :username="user.username" @changeUsername="changeUsername($event)"/>

export default {
...
methods: {
   changeUsername (username) {
     this.user.username = username;
   }
}
}