we add a grocerylist to the data of todolist component and pass that data to the child component todoitem as props by looping through that array. in todoitem we add the props object with an array and the name of the prop that we pass and now we have access to that prop and its properties.
This is a contrived example, but we have managed to separate our app into two smaller units, and the child is reasonably well-decoupled from the parent via the props interface. We can now further improve our <todo-item> component with more complex template and logic without affecting the parent app.

In a large application, it is necessary to divide the whole app into components to make development manageable